requirements.txt
pytest==6.2.2
PyMySQL==0.10.1
allure-pytest==2.8.34
requests==2.25.1
PyYAML==5.3.1

下载不动了 用镜像  pip install pytest -i https://mirrors.aliyun.com/pypi/simple/